refactor: 1주차 2번째 리뷰 반영
