+++ BE/src/main/java/yanse/airbnb/domain/image/Image.java
@@ -1,14 +1,17 @@
-package yanse.airbnb.domain.main;
+package yanse.airbnb.domain.image;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import yanse.airbnb.type.ImageType;
 
 @Entity
-public class MainImage {
+public class Image {
 
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(name = "image_id")
 	private Long id;
 
 	private String url;
