+++ BE/src/main/java/yanse/airbnb/domain/reservation/DetailFee.java
@@ -1,11 +1,11 @@
 package yanse.airbnb.domain.reservation;
 
 import javax.persistence.Embeddable;
+import yanse.airbnb.type.DiscountPolicy;
 
 @Embeddable
 public class DetailFee {
 
-	private int night;
 	private DiscountPolicy discountPolicy;
 	private double cleaningFee;
 	private double serviceTax;
