+++ BE/src/main/java/yanse/airbnb/domain/room/Room.java
@@ -15,8 +15,6 @@ public class Room {
 	@Column(name = "room_id")
 	private Long id;
 
-	private int totalRoom;
-
 	private int price;
 
 	private String name;
