+++ BE/src/main/java/yanse/airbnb/domain/wish/Wish.java
@@ -9,16 +9,15 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToOne;
 import yanse.airbnb.domain.member.Member;
 import yanse.airbnb.domain.room.Room;
 
 
 @Entity
-public class WishList {
+public class Wish {
 
 	@Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(name = "wishlist_id")
+	@Column(name = "wish_id")
 	private Long id;
 
 	@ManyToOne(fetch = LAZY)
