+++ BE/src/main/java/yanse/airbnb/type/RoomType.java
@@ -1,4 +1,4 @@
-package yanse.airbnb.domain.room;
+package yanse.airbnb.type;
 
 import lombok.RequiredArgsConstructor;
